requirements.txt
h5py
easydict==1.7
graphviz==0.8.4
gym==0.10.5
imageio==2.3.0
matplotlib==2.2.2
numpy==1.15.4
Pillow==5.2.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
tensorboardX==1.2
torch==1.4.0
torchvision==0.4.1
tqdm==4.23.4
nibabel==2.5.0
pandas
opencv-python
paramiko
polyaxon-client==0.5.6
pathlib
SimpleITK==1.2.4
PyYAML
albumentations
medpy

